requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
librosa>=0.10.0
numpy>=1.21.0
nltk>=3.6.3
pydub>=0.25.1
ffmpeg-python>=0.2.0
deep-translator>=1.11.4
openai-whisper>=20231117
torch>=2.0.0